refactor to separate calcBreathMark into a separate function
