+++ src/export/mnx_mapping.cpp
@@ -88,185 +88,6 @@ JumpType convertTextToJump(const std::string& text, const std::optional<std::str
     return JumpType::None;
 }
 
-std::vector<EventMarkingType> calcMarkingType(
-    const details::ArticulationAssign::SelectedSymbolContext& articContext,
-    std::optional<int>& numMarks,
-    std::optional<mnx::BreathMarkSymbol>& breathMark)
-{
-    auto checkSymbol = [&](char32_t sym, const MusxInstance<FontInfo>& fontInfo) -> std::vector<EventMarkingType> {
-        // First, check for standard Unicode chars
-        switch (sym) {
-            case 0x1D167:
-            case 0x1D16A:
-                numMarks = 1;
-                return { EventMarkingType::Tremolo };
-            case 0x1D168:
-            case 0x1D16B:
-                numMarks = 2;
-                return { EventMarkingType::Tremolo };
-            case 0x1D169:
-            case 0x1D16C:
-                numMarks = 3;
-                return { EventMarkingType::Tremolo };
-            case 0x1D17B: return { EventMarkingType::Accent };
-            case 0x1D17C: return { EventMarkingType::Staccato };
-            case 0x1D17D: return { EventMarkingType::Tenuto };
-            case 0x1D17E: return { EventMarkingType::Staccatissimo };
-            case 0x1D17F: return { EventMarkingType::StrongAccent };
-            default: break;                
-        }
-
-        /// @todo Spiccato marking has no SMuFl glyph!
-        if (auto glyphName = utils::smuflGlyphNameForFont(fontInfo, sym)) {
-            if (glyphName == "articAccentAbove" || glyphName == "articAccentBelow" || glyphName == "articAccentAboveLegacy") {
-                return { EventMarkingType::Accent };
-            } else if (glyphName == "articAccentStaccatoAbove" || glyphName == "articAccentStaccatoBelow" ||
-                    glyphName == "articAccentStaccatoAboveLegacy" || glyphName == "articAccentStaccatoBelowLegacy") {
-                return { EventMarkingType::Accent, EventMarkingType::Staccato };
-            } else if (glyphName == "articTenutoAccentAbove" || glyphName == "articTenutoAccentBelow" ||
-                    glyphName == "articTenutoAccentAboveLegacy" || glyphName == "articTenutoAccentBelowLegacy") {
-                return { EventMarkingType::Accent, EventMarkingType::Tenuto };
-            } else if (glyphName == "stringsDownBow" || glyphName == "stringsDownBowTurned") {
-                return { EventMarkingType::BowDirectionDown };
-            } else if (glyphName == "stringsUpBow" || glyphName == "stringsUpBowTurned") {
-                return { EventMarkingType::BowDirectionUp };
-            } else if (glyphName == "breathMarkComma" || glyphName == "breathMarkCommaLegacy") {
-                breathMark = mnx::BreathMarkSymbol::Comma;
-                return { EventMarkingType::Breath };
-            } else if (glyphName == "breathMarkTick") {
-                breathMark = mnx::BreathMarkSymbol::Tick;
-                return { EventMarkingType::Breath };
-            } else if (glyphName == "breathMarkUpbow") {
-                breathMark = mnx::BreathMarkSymbol::Upbow;
-                return { EventMarkingType::Breath };
-            } else if (glyphName == "breathMarkSalzedo") {
-                breathMark = mnx::BreathMarkSymbol::Salzedo;
-                return { EventMarkingType::Breath };
-            } else if (glyphName == "articStaccatissimoAbove" || glyphName == "articStaccatissimoBelow") {
-                return { EventMarkingType::Spiccato };
-            } else if (glyphName == "articStaccatissimoStrokeAbove" || glyphName == "articStaccatissimoStrokeBelow" ||
-                    glyphName == "articStaccatissimoWedgeAbove" || glyphName == "articStaccatissimoWedgeBelow") {
-                return { EventMarkingType::Staccatissimo };
-            } else if (glyphName == "articStaccatoAbove" || glyphName == "articStaccatoBelow") {
-                return { EventMarkingType::Staccato };
-            } else if (glyphName == "articMarcatoStaccatoAbove" || glyphName == "articMarcatoStaccatoBelow" ||
-                    glyphName == "articMarcatoStaccatoAboveLegacy" || glyphName == "articMarcatoStaccatoBelowLegacy") {
-                return { EventMarkingType::Staccato, EventMarkingType::StrongAccent };
-            } else if (glyphName == "articTenutoStaccatoAbove" || glyphName == "articTenutoStaccatoBelow" ||
-                    glyphName == "articTenutoStaccatoAboveLegacy" || glyphName == "articTenutoStaccatoBelowLegacy") {
-                return { EventMarkingType::Staccato, EventMarkingType::Tenuto };
-            } else if (glyphName == "articStressAbove" || glyphName == "articStressBelow") {
-                return { EventMarkingType::Stress };
-            } else if (glyphName == "articUnstressAbove" || glyphName == "articUnstressBelow") {
-                return { EventMarkingType::Unstress };
-            } else if (glyphName == "articMarcatoAbove" || glyphName == "articMarcatoBelow") {
-                return { EventMarkingType::StrongAccent };
-            } else if (glyphName == "articMarcatoTenutoAbove" || glyphName == "articMarcatoTenutoBelow") {
-                return { EventMarkingType::StrongAccent, EventMarkingType::Tenuto };
-            } else if (glyphName == "articTenutoAbove" || glyphName == "articTenutoBelow") {
-                return { EventMarkingType::Tenuto };
-            } else if (glyphName == "stemPendereckiTremolo" || glyphName == "buzzRoll" ||
-                    glyphName == "pendereckiTremolo" || glyphName == "unmeasuredTremolo" ||
-                    glyphName == "unmeasuredTremoloSimple" || glyphName == "stockhausenTremolo") {
-                numMarks = 0;
-                return { EventMarkingType::Tremolo };
-            } else if (glyphName == "tremolo1" || glyphName == "tremoloFingered1" || glyphName == "tremolo1Alt") {
-                numMarks = 1;
-                return { EventMarkingType::Tremolo };
-            } else if (glyphName == "tremolo2" || glyphName == "tremoloFingered2" || glyphName == "tremolo2Alt") {
-                numMarks = 2;
-                return { EventMarkingType::Tremolo };
-            } else if (glyphName == "tremolo3" || glyphName == "tremoloFingered3" || glyphName == "tremolo3Alt") {
-                numMarks = 3;
-                return { EventMarkingType::Tremolo };
-            } else if (glyphName == "tremolo4" || glyphName == "tremoloFingered4" || glyphName == "tremolo4Legacy") {
-                numMarks = 4;
-                return { EventMarkingType::Tremolo };
-            } else if (glyphName == "tremolo5" || glyphName == "tremoloFingered5" || glyphName == "tremolo5Legacy") {
-                numMarks = 5;
-                return { EventMarkingType::Tremolo };
-            }
-        }
-        return {};
-    };
-    
-    auto checkShape = [&](Cmper shapeId) -> std::vector<EventMarkingType> {
-        const auto& articDef = *articContext.definition;
-        if (auto shape = articDef.getDocument()->getOthers()->get<others::ShapeDef>(articDef.getRequestedPartId(), shapeId)) {
-            switch (shape->recognize()) {
-            case KnownShapeDefType::TenutoMark:
-                return { EventMarkingType::Tenuto };
-            default:
-                break;
-            }
-        }
-        return {};
-    };
-
-    auto result = articContext.symbol.isShape
-                ? checkShape(articContext.symbol.shapeId)
-                : checkSymbol(articContext.symbol.character, articContext.symbol.font);
-    return result;
-}
-
-std::optional<mnx::Fermata> calcFermata(const MusxInstance<FontInfo>& fontInfo, char32_t sym, VerticalPlacement placement)
-{
-    if (placement == VerticalPlacement::NotApplicable) {
-        return std::nullopt;
-    }
-
-    std::optional<mnx::Fermata> result;
-
-    if (auto glyphName = utils::smuflGlyphNameForFont(fontInfo, sym)) {
-        auto calcOrient = [&]() {
-            auto orient = enumConvert<mnx::Orientation>(placement);
-            if (orient == mnx::Orientation::Auto) {
-                if (glyphName->size() >= 5 && glyphName->compare(glyphName->size() - 5, 5, "Above") == 0) {
-                    return mnx::Orientation::Above;
-                }
-                if (glyphName->size() >= 5 && glyphName->compare(glyphName->size() - 5, 5, "Below") == 0) {
-                    return mnx::Orientation::Below;
-                }
-            }
-            return orient;
-        };
-        auto makeFermata = [&](mnx::FermataSymbol symbol, mnx::FermataDuration duration = mnx::FermataDuration::Auto) {
-            auto& fermata = result.emplace();
-            fermata.set_or_clear_symbol(symbol);
-            fermata.set_or_clear_duration(duration);
-            fermata.set_or_clear_orient(calcOrient());
-        };
-
-        if (glyphName == "fermataAbove" || glyphName == "fermataBelow") {
-            makeFermata(mnx::FermataSymbol::Normal);
-        } else if (glyphName == "fermataVeryShortAbove" || glyphName == "fermataVeryShortBelow") {
-            makeFermata(mnx::FermataSymbol::DoubleAngled, mnx::FermataDuration::VeryShort);
-        } else if (glyphName == "fermataShortAbove" || glyphName == "fermataShortBelow") {
-            makeFermata(mnx::FermataSymbol::Angled, mnx::FermataDuration::Short);
-        } else if (glyphName == "fermataLongAbove" || glyphName == "fermataLongBelow") {
-            makeFermata(mnx::FermataSymbol::Square, mnx::FermataDuration::Long);
-        } else if (glyphName == "fermataVeryLongAbove" || glyphName == "fermataVeryLongBelow") {
-            makeFermata(mnx::FermataSymbol::DoubleSquare, mnx::FermataDuration::VeryLong);
-        } else if (glyphName == "fermataLongHenzeAbove" || glyphName == "fermataLongHenzeBelow") {
-            makeFermata(mnx::FermataSymbol::DoubleDot, mnx::FermataDuration::Long);
-        } else if (glyphName == "fermataShortHenzeAbove" || glyphName == "fermataShortHenzeBelow") {
-            makeFermata(mnx::FermataSymbol::HalfCurve, mnx::FermataDuration::Short);
-        } else if (glyphName == "curlewSign") {
-            makeFermata(mnx::FermataSymbol::Curlew);
-        }
-    }
-
-    return result;
-}
-
-std::optional<mnx::Fermata> calcFermata(const MusxInstance<FontInfo>& fontInfo, const std::string& symStr, VerticalPlacement placement)
-{
-    if (const auto sym = utils::utf8ToCodepoint(symStr)) {
-        return calcFermata(fontInfo, sym.value(), placement);
-    }
-    return std::nullopt;
-}
-
 mnx::NoteValue::Required mnxNoteValueFromEdu(Edu duration)
 {
     auto [base, dots] = calcDurationInfoFromEdu(duration);
